requirements.txt
# requirements.txt

# Core LangChain
langchain>=0.1.0
langchain-community>=0.0.30
langchain-core>=0.1.51
langchain-ollama>=0.1.0

# Vector Store
faiss-cpu>=1.7.0

# # Embeddings
# sentence-transformers>=2.2.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0

# JSON processing
pydantic>=2.0.0

# Testing
pytest>=7.0.0

# MLOps & Experiment Tracking
mlflow==2.9.0

# Web Framework & API
fastapi==0.104.1
uvicorn==0.24.0

# Logging
python-json-logger>=2.0.0


# 
# # Core LLM & Language Model Framework
# langchain==0.2.11
# langchain-core==0.2.27
# langchain-anthropic==0.1.21
# anthropic>=0.21.0
# 
# # Models
# openai==1.3.0
# ollama==0.1.6
# 
# # Web Framework & API
# fastapi==0.104.1
# uvicorn==0.24.0
# 
# # Configuration & Environment
# pydantic==2.5.0
# python-dotenv==1.0.0
# setuptools>=68.0.0
# 
# # MLOps & Experiment Tracking
# mlflow==2.9.0
# 
# # Templating & Text Processing
# jinja2==3.1.2
# sentence-transformers>=2.2.0
# 
# # Testing
# pytest==7.4.3
# 
# # HTTP & Networking
# requests==2.31.0
